feat: 🎸 add missing callback API methods and some sycn API ones
